backend/system/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
whitenoise==6.11.0
drf-spectacular>=0.27.0